Moved auth.service.ts to app\core\services, set new destination of this file and removed unused import
