+++ src/app/features/auth/auth.module.ts
@@ -12,7 +12,7 @@ import {MatToolbarModule} from "@angular/material/toolbar";
 import {MatIconModule} from "@angular/material/icon";
 import {MatButtonModule} from "@angular/material/button";
 import {FormsModule} from "@angular/forms";
-import {AuthService} from "./auth.service";
+import {AuthService} from "../../core/services/auth.service";
 import { RegisterComponent } from './register/register.component';
 import {MatProgressBarModule} from "@angular/material/progress-bar";
