+++ src/app/features/auth/login/login.component.ts
@@ -1,6 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import {AuthService} from "../auth.service";
-import {catchError} from "rxjs";
+import {AuthService} from "../../../core/services/auth.service";
 
 @Component({
   selector: 'app-login',
