+++ src/app/features/auth/register/register.component.ts
@@ -1,5 +1,5 @@
 import {Component, OnInit} from '@angular/core';
-import {AuthService} from "../auth.service";
+import {AuthService} from "../../../core/services/auth.service";
 
 @Component({
   selector: 'app-register',
